removing OtlpTransportFactoryInterface
