+++ examples/traces/features/exporters/otlp_http.php
@@ -5,15 +5,15 @@ require __DIR__ . '/../../../../vendor/autoload.php';
 
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
+use OpenTelemetry\Contrib\Otlp\OtlpHttpTransportFactory;
 use OpenTelemetry\Contrib\Otlp\SpanExporter;
-use OpenTelemetry\Contrib\OtlpHttp\OtlpHttpTransportFactory;
 use OpenTelemetry\SDK\Common\Log\LoggerHolder;
 use OpenTelemetry\SDK\Trace\SpanProcessor\SimpleSpanProcessor;
 use OpenTelemetry\SDK\Trace\TracerProvider;
 
 LoggerHolder::set(new Logger('otlp-example', [new StreamHandler('php://stderr')]));
 
-$transport = (new OtlpHttpTransportFactory())->create('http://collector:4318');
+$transport = (new OtlpHttpTransportFactory())->create('http://collector:4318/v1/traces', 'application/x-protobuf');
 $exporter = new SpanExporter($transport);
 
 echo 'Starting OTLP example';
